move GradlePropertiesFileSource to separate class
